+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 - Add `as_dict()` utility method to `Score`
+- Add `token_limit` option for limiting the number of tokens used per sample.
+- Add `message_limit` option for limiting the number of messages per sample (replaces deprecated `max_messages` option),
 
 ## v0.3.40 (6 October 2024)